Add preview encoder, ffmpeg helpers, and preview task
Introduce end-to-end support for generating sketch preview clips and shared FFmpeg helpers. Adds:

- src/lib/createSketchPreviews.ts: new task that captures frames from templates via a headless browser, writes temporary frames, and encodes a short WebM preview (2.5s @ 24fps) per sketch. Updates templates/metadata.json to mark generated previews.
- src/lib/encodePreviewFromFrames.ts: encodes PNG frame sequences into VP9 WebM previews (CRF 37, cpu-used 4), using the concat demuxer and a generated per-frame list file.
- src/utils/ffmpegFrameHelpers.ts: shared utilities for natural filename sorting, escaping paths for ffmpeg concat, and writing concat list files.
- src/utils/fileExists.ts: small fs.exists helper.

Also refactors existing code to reuse the new helpers:

- src/lib/createSketchThumbnails.ts now uses fileExists instead of a local exists implementation.
- src/utils/encodeVideoFromFrames.ts now uses listPngFramesSorted and writeConcatList from ffmpegFrameHelpers instead of duplicating that logic.

These changes centralize frame-list/ordering logic, add a dedicated preview encoder, and reduce duplication.
