+++ src/utils/fileExists.ts
@@ -1,0 +1,10 @@
+import fs from "node:fs/promises";
+
+export default async function fileExists( filePath: string ) {
+  try {
+    await fs.stat( filePath );
+    return true;
+  } catch {
+    return false;
+  }
+}
